Added linear Schrodinger
lse.py solves the linear one-particle Schrodinger equation. The user can currently choose a free particle, or the Poschl-Teller/sech^2 potential, by un-commenting the relevant section of code. The default outputs are temperature plots of the wavefunction (real and imaginary parts separately) and a plot of the error in L^2 norm (since linear Schrodinger solutions should keep this constant!).